Move enonic-connection-config to xpAdapter folder instead of src root
