+++ src/customXp/partSelector.ts
@@ -1,5 +1,5 @@
 // APP_NAME helps to fully qualify XP-component name strings in the connected XP app:
-import {APP_NAME} from "../enonic-connection-config";
+import {APP_NAME} from "../xpAdapter/enonic-connection-config";
 
 import {ReactView} from "./contentSelector";
