+++ src/customXp/parts/Footer.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import {APP_NAME} from "../../enonic-connection-config";
+import {APP_NAME} from "../../xpAdapter/enonic-connection-config";
 import Footer from "../../components/blocks/Footer";
 
 // fully qualified XP part name:
