+++ src/customXp/parts/PersonList.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import {APP_NAME} from "../../enonic-connection-config";
+import {APP_NAME} from "../../xpAdapter/enonic-connection-config";
 
 // fully qualified XP part name:
 export const PERSONLIST_PART_NAME = `${APP_NAME}:personList`;
