+++ src/xpAdapter/views/_BaseContent.tsx
@@ -7,7 +7,7 @@ import DefaultContentView from "../../customXp/contentTypes/_DefaultView";
 import contentSelector, {ContentSelection} from "../../customXp/contentSelector";
 
 
-const BaseContent =(props: FetchContentResult) => {
+const BaseContent = (props: FetchContentResult) => {
     const {content, meta, page} = props;
 
     if (!content) {
